Added empty basic methods to the classes (#4)
* Changed data type of the variables from BigDecimal to Double

* Added increase/reduce methods to the DaniloBank class

* Added increase/reduce methods for the balance variable

* Added payCredit function

* Added reduce total customers credit function

* Changes some comments

+++ src/danylo/bank/app/customers/BronzeCustomer.java
@@ -2,7 +2,6 @@ package danylo.bank.app.customers;
 
 import danylo.bank.app.customers.interfaces.BronzeCustomerInterface;
 
-import java.math.BigDecimal;
 
 /**
  * Is subclass for the Customer abstract class.
@@ -20,17 +19,17 @@ public class BronzeCustomer extends Customer implements BronzeCustomerInterface 
      * @param customerID     Integer representing the ID of the customer
      * @param salaryPerMonth BigDecimal representing the salary per month of the customer
      */
-    public BronzeCustomer(String customerName, Integer customerID, BigDecimal salaryPerMonth) {
+    public BronzeCustomer(String customerName, Integer customerID, Double salaryPerMonth) {
         super(customerName, customerID, salaryPerMonth);
     }
 
     @Override
-    public void topUpBalance(BigDecimal amount) {
+    public void topUpBalance(Double amount) {
 
     }
 
     @Override
-    public void withdrawMoney(BigDecimal amount) {
+    public void withdrawMoney(Double amount) {
 
     }
 }
