+++ src/danylo/bank/app/customers/interfaces/BronzeCustomerInterface.java
@@ -1,6 +1,5 @@
 package danylo.bank.app.customers.interfaces;
 
-import java.math.BigDecimal;
 
 /**
  * This interface responsible for keeping bronze customer methods.
@@ -12,7 +11,7 @@ public interface BronzeCustomerInterface {
      *
      * @param amount The value by which balance and bank account will be increased
      */
-    void topUpBalance(BigDecimal amount);
+    void topUpBalance(Double amount);
 
     /**
      * This method withdraws money from customer's balance and reduces the total bank account
@@ -20,5 +19,5 @@ public interface BronzeCustomerInterface {
      *
      * @param amount The value by which balance and bank account will be reduced
      */
-    void withdrawMoney(BigDecimal amount);
+    void withdrawMoney(Double amount);
 }
